feat: add scan management on stock moves (#1169)
* RM#105014
