+++ changelogs/unreleased/105014.json
@@ -1,0 +1,5 @@
+{
+  "title": "StockMove: add scan management",
+  "type": "feat",
+  "packages": "stock"
+}
